xml doc formattering
Former-commit-id: 648256bdcd8bdcb892a9a64649fef46fb62d7d0d

+++ SpaceComponentSystem/Systems/DropSystem.cs
@@ -12,16 +12,14 @@ using Space.ComponentSystem.Messages;
 
 namespace Space.ComponentSystem.Systems
 {
-    /// <summary>
-    /// Manages item drops by reacting to death events.
-    /// </summary>
+    /// <summary>Manages item drops by reacting to death events.</summary>
     public sealed class DropSystem : AbstractSystem, IMessagingSystem, IUpdatingSystem
     {
         #region Fields
 
         /// <summary>
-        /// List of drops to sample when we update. This is accumulated from
-        /// death events, to allow thread safe sampling in one go.
+        ///     List of drops to sample when we update. This is accumulated from death events, to allow thread safe sampling
+        ///     in one go.
         /// </summary>
         [CopyIgnore, PacketizerIgnore]
         private List<Tuple<string, FarPosition>> _dropsToSample = new List<Tuple<string, FarPosition>>();
@@ -30,9 +28,7 @@ namespace Space.ComponentSystem.Systems
 
         #region Logic
 
-        /// <summary>
-        /// Removes entities that died this frame from the manager.
-        /// </summary>
+        /// <summary>Removes entities that died this frame from the manager.</summary>
         /// <param name="frame">The current simulation frame.</param>
         public void Update(long frame)
         {
@@ -43,11 +39,10 @@ namespace Space.ComponentSystem.Systems
             }
             _dropsToSample.Clear();
         }
-        
+
         /// <summary>
-        /// Queues a drops for one or more items from the specified item pool at the
-        /// specified position. The actual drop will be performed when the drop system
-        /// updates.
+        ///     Queues a drops for one or more items from the specified item pool at the specified position. The actual drop
+        ///     will be performed when the drop system updates.
         /// </summary>
         /// <param name="poolName">The name of the item pool to sample from.</param>
         /// <param name="position">The position at which to drop the items.</param>
@@ -59,9 +54,7 @@ namespace Space.ComponentSystem.Systems
             }
         }
 
-        /// <summary>
-        /// Performs the actual drop sampling.
-        /// </summary>
+        /// <summary>Performs the actual drop sampling.</summary>
         /// <param name="poolName">The name of the item pool to sample from.</param>
         /// <param name="position">The position at which to drop the items.</param>
         private void SampleDrop(string poolName, FarPosition position)
@@ -71,14 +64,14 @@ namespace Space.ComponentSystem.Systems
 
             // Get the list of possible drops.
             var dropInfo = new List<ItemPool.DropInfo>(pool.Items);
-            
+
             // Randomizer used for sampling of items. Seed it based on the item
             // pool and the drop position, to get a deterministic result for
             // each drop, regardless in which order they happen.
             var hasher = new Hasher();
             hasher.Write(poolName);
             hasher.Write(position);
-            var random = new MersenneTwister((ulong)hasher.Value);
+            var random = new MersenneTwister(hasher.Value);
 
             // And shuffle it. This is important, to give each entry an equal
             // chance to be picked. Otherwise the first few entries have a much
@@ -106,7 +99,7 @@ namespace Space.ComponentSystem.Systems
                     var entity = FactoryLibrary.SampleItem(Manager, item.ItemName, position, random);
 
                     // Make the item visible.
-                    var renderer = ((TextureRenderer)Manager.GetComponent(entity, TextureRenderer.TypeId));
+                    var renderer = ((TextureRenderer) Manager.GetComponent(entity, TextureRenderer.TypeId));
                     if (renderer != null)
                     {
                         renderer.Enabled = true;
@@ -122,9 +115,7 @@ namespace Space.ComponentSystem.Systems
             }
         }
 
-        /// <summary>
-        /// Drop items when entities die.
-        /// </summary>
+        /// <summary>Drop items when entities die.</summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="message">The message.</param>
         public void Receive<T>(T message) where T : struct
@@ -137,10 +128,10 @@ namespace Space.ComponentSystem.Systems
 
             var entity = cm.Value.KilledEntity;
 
-            var drops = ((Drops)Manager.GetComponent(entity, Drops.TypeId));
+            var drops = ((Drops) Manager.GetComponent(entity, Drops.TypeId));
             if (drops != null)
             {
-                var translation = ((Transform)Manager.GetComponent(entity, Transform.TypeId)).Translation;
+                var translation = ((Transform) Manager.GetComponent(entity, Transform.TypeId)).Translation;
                 Drop(drops.ItemPool, ref translation);
             }
         }
@@ -150,18 +141,13 @@ namespace Space.ComponentSystem.Systems
         #region Copying
 
         /// <summary>
-        /// Servers as a copy constructor that returns a new instance of the same
-        /// type that is freshly initialized.
-        /// 
-        /// <para>
-        /// This takes care of duplicating reference types to a new copy of that
-        /// type (e.g. collections).
-        /// </para>
+        ///     Servers as a copy constructor that returns a new instance of the same type that is freshly initialized.
+        ///     <para>This takes care of duplicating reference types to a new copy of that type (e.g. collections).</para>
         /// </summary>
         /// <returns>A cleared copy of this system.</returns>
         public override AbstractSystem NewInstance()
         {
-            var copy = (DropSystem)base.NewInstance();
+            var copy = (DropSystem) base.NewInstance();
 
             copy._dropsToSample = new List<Tuple<string, FarPosition>>();
 
@@ -169,17 +155,17 @@ namespace Space.ComponentSystem.Systems
         }
 
         /// <summary>
-        /// Creates a deep copy of the system. The passed system must be of the
-        /// same type.
-        /// <para>
-        /// This clones any contained data types to return an instance that
-        /// represents a complete copy of the one passed in.
-        /// </para>
+        ///     Creates a deep copy of the system. The passed system must be of the same type.
+        ///     <para>
+        ///         This clones any contained data types to return an instance that represents a complete copy of the one passed
+        ///         in.
+        ///     </para>
         /// </summary>
         /// <param name="into">The instance to copy into.</param>
         public override void CopyInto(AbstractSystem into)
         {
-            System.Diagnostics.Debug.Assert(_dropsToSample.Count == 0, "Drop system got drop requests after its update, but before copying.");
+            System.Diagnostics.Debug.Assert(
+                _dropsToSample.Count == 0, "Drop system got drop requests after its update, but before copying.");
 
             base.CopyInto(into);
         }
